DIstributed version of the pipeline workflows
This new framework submits CBDA-SL learner steps/iterations as single jobs.
All the R scripts and txt files (e.g., Input arguments and Dataset files) are located locally (remember to edit the pipe file with the correct locations on your machine).
This new version includes bartMachine as part of the Superlearner library. The CV.Superlearner is commented out in the SUperlearner Loop code due to high computational cost. We are investigating ways to make it faster. Right now each job has a run time of 2-3 minutes.
The MSE and Feature Mining module ranks the top features within single experiments (i.e., input i-exp) and across all of them.
In the R script of the Superlearner loop, 2 blocks are commented out: i) generation of a random/null dataset based on the Input Dataset  (Lines 47-58), and ii) knockoff filter implementation (Lines 375-383).
I am working on generating a fake dataset where we can test the knockoff filter and use it to guide MSE and feature minin