Updated pipe file and R scripts
The pipe file runs on Cranium and it sequentially performs 30 experiments (flow control mechanisms in place). Each experiment is 3000 independent jobs. The specs of the 30 experiments are listed in the text file arg_3000_30.txt (each row is an experiment's spec).
The dataset used is "fake", built ad hoc to test the feature mining capabilities of both the CBDA-SL and knockoff filter algorithms. It's called binomial because the outcome of interest is binomial. The master code is in the CBDA R script (CBDA_CV_SL_master_V14_distributed.R). The post optimization script (run after all the 30 modules have completed) is in Data_Consolidation_MSE_FeatureMining_V14.R.
The final product is a set of 30 RData workspaces that are then processed to obtain 60 histograms (2 for each experiment) with relative frequencies of the 100 features in the dataset (spikes are the ones to spot). The file to process the RData is being implemented as R Markdown (it will be pushed shortly).
.