UPDATED VERSION OF THE PIPELINE WORKFLOW FILES
The pipe file runs on Cranium and it sequentially performs 12 experiments. Each experiment is 9000 independent jobs. The specs of the 12 experiments are listed in the text file arg_Dec9.txt (each row is an experiment's spec). Due to the max job submission on Cranium (i.e., 3000), each experiment has 3 modules of 3000 jobs each. Flow control mechanisms are in place between each experiment's module, as well as between different experiments.
The dataset used is "fake", built ad hoc to test the feature mining capabilities of both the CBDA-SL and knockoff filter algorithms. It's called binomial because the outcome of interest is binomial. The master code is in the CBDA R script (CBDA_CV_SL_master_V15_distributed.R). The post optimization script (run after each experiment module has completed) is in Data_Consolidation_MSE_FeatureMining_V14.R.
The final product is a set of 12 RData workspaces that are then processed to obtain 24 histograms (2 for each experiment) wit